RDKC-15902: Re-Structure rfc xconf handlers for rdkc
